requirements_file.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==23.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.8
dp-accounting==0.4.3
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-vizier==0.1.11
googleapis-common-protos==1.61.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.62.3
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.14
jaxlib==0.4.14
Jinja2==3.1.6
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
ml-dtypes==0.2.0
mpmath==1.3.0
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==22.0
pandas==2.2.3
pillow==11.1.0
pip==23.0.1
portpicker==1.6.0
protobuf==4.25.6
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.9.3
seaborn==0.13.2
setuptools==65.5.0
six==1.17.0
SQLAlchemy==1.4.20
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow_federated==0.87.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-model-optimization==0.7.5
tensorflow_privacy==0.9.0
tensorflow-probability==0.22.1
termcolor==3.0.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
